feat: complete Phase 3 rich bookmark implementation
- Add comprehensive test suite for bookmark functionality
- Move all bookmark-related tests to bookmark.rs module
- Clean up converter.rs by removing bookmark tests and reducing pub use
- Improve code organization and module boundaries
- All 57 tests passing with zero warnings
